tweak: unselectable row
